moved create participant code around in order to bump up code climate stats
