+++ babel.config.js
@@ -7,5 +7,6 @@ module.exports = {
 	plugins: [
 		"@babel/plugin-transform-modules-commonjs",
 		"@babel/plugin-proposal-class-properties",
+		"@babel/plugin-transform-runtime",
 	],
 };
