+++ client/package.json
@@ -1,41 +1,44 @@
 {
-  "name": "secret-santa-app",
-  "version": "1.0.0",
-  "private": true,
-  "proxy": "http://localhost:3001/",
-  "dependencies": {
-    "@ant-design/icons": "^4.2.2",
-    "@auth0/auth0-react": "^1.0.0",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "ant-design-pro": "^2.3.2",
-    "antd": "^4.6.4",
-    "axios": "^0.19.2",
-    "dotenv": "^8.2.0",
-    "graphql": "^15.3.0",
-    "graphql-request": "^3.1.0",
-    "moment": "^2.29.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-router-dom": "^5.0.1",
-    "react-scripts": "^3.0.1"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "eject": "react-scripts eject"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  }
+	"name": "secret-santa-app",
+	"version": "1.0.0",
+	"private": true,
+	"proxy": "http://localhost:3001/",
+	"dependencies": {
+		"@ant-design/icons": "^4.2.2",
+		"@auth0/auth0-react": "^1.0.0",
+		"@testing-library/jest-dom": "^4.2.4",
+		"@testing-library/react": "^9.5.0",
+		"@testing-library/user-event": "^7.2.1",
+		"ant-design-pro": "^2.3.2",
+		"antd": "^4.6.4",
+		"axios": "^0.19.2",
+		"dotenv": "^8.2.0",
+		"graphql": "^15.3.0",
+		"graphql-request": "^3.1.0",
+		"moment": "^2.29.0",
+		"react": "^16.8.6",
+		"react-dom": "^16.8.6",
+		"react-router-dom": "^5.0.1",
+		"react-scripts": "^3.0.1"
+	},
+	"scripts": {
+		"start": "react-scripts start",
+		"build": "react-scripts build",
+		"eject": "react-scripts eject"
+	},
+	"browserslist": {
+		"production": [
+			">0.2%",
+			"not dead",
+			"not op_mini all"
+		],
+		"development": [
+			"last 1 chrome version",
+			"last 1 firefox version",
+			"last 1 safari version"
+		]
+	},
+	"devDependencies": {
+		"@babel/plugin-transform-runtime": "^7.11.5"
+	}
 }
